Version 1.28
  * Added a machineenv command.
  * More BSD services: make-utx-log, openbsd-cron
  * More cross-platform services: kmod@fuse, mogstored, mogilefsd, kamailio, NetworkManager-dispatcher, clamd, pcscd
  * In ideal mode, convert-systemd-services defaults to always restarting as per daemontools family norms.
  * In quirks mode, convert-systemd-services implements more of systemd's standard I/O redirection semantics.
  * convert-systemd-services now translates some Linux-idiosyncratic limits.
  * SoftTerm no longer sideways-scrolls in response to CUF and CUB.
  * MySQL and MariaDB services are now instantiated by importing from the my.cnf configuration information.
  * Fixed a bug where /usr on a separate volume to / would prevent package installation on BSD.
  * All of the FreeDesktop Kit service bundles now have aliases for use with Desktop Bus activation.
  * There is now a very simple dbus-daemon-launch-helper shim that devolves to invoking service management.
  * The supervise/ directory control API now supports '_', which is like 'd' but does not generate TERM or CONT signals.
  * svc -x is now implemented.
  * gnucron is now debian-cron, as GNU is a slight misnomer; it being a GPL Debian project not an FSF GNU project.
  * ulimit and softlimit now support SI and IEC suffixes for explicit scale factors (when the implicit scale factor is 1).
  * An experimental extension has been added to the supervise/status file that records the exit status of the last starting/running main process to terminate.
  * There are now a move-to-control-group and a delegate-control-group-to command.
  * There is now a tcpserver shim, for use with FreeBSD/PC-BSD ports of publicfile et al..
  * Each user now has an exit target for the per-user-manager and thus the latter shuts down cleanly.
  * The OBOEs in console-terminal-emulator's SCO function key handling have been fixed.
  * SoftTerm now supports some of the more esoteric old FreeBSD kernel terminal emulator's control sequences, for increased cons25 compatibility.
  * The -s boot option in BSD now enters emergency mode not rescue mode.
  * console-terminal-emulator now has a mechanism for "bracketed paste" input.
  * system-control preset --ttys now defaults to disable if /etc/ttys is absent.
  * rcctl is now in an openbsd-shims package, which is not installed on OpenBSD.
  * The service manager can now be run under OpenBSD rc.
  * The various -listen tools can now merge into an existing set of systemd LISTEN_FDS file descriptors; and convert-systemd-units makes use of this for multi-listener sockets.

+++ source/convert/dnscache.do
@@ -7,21 +7,60 @@
 # This is invoked by general-services.do .
 #
 
-s="`system-control find dnscache`"
-
 set_if_unset() { if test -z "`system-control print-service-env \"$1\" \"$2\"`" ; then system-control set-service-env "$1" "$2" "$3" ; echo "$s: Defaulted $2 to $3." ; fi ; }
 dir_not_empty() { test -n "`/bin/ls -A \"$1\"`" ; }
 
-redo-ifchange general-services
+# These get us *only* the configuration variables, safely.
+read_rc() { clearenv read-conf rc.conf "`which printenv`" "$1" ; }
+list_network_addresses() { read_rc network_addresses || echo 127.0.0.1 | while read -r i ; do echo "$i" ; done ; }
 
-set_if_unset dnscache IPSEND 0.0.0.0
-set_if_unset dnscache IP 127.0.0.1
-set_if_unset dnscache DATALIMIT 3000000
-set_if_unset dnscache CACHESIZE 1000000
-set_if_unset dnscache ROOT "$s/service/root"
+redo-ifchange rc.conf general-services "dnscache@.socket" "dnscache.service"
 
-system-control print-service-env dnscache >> "$3"
+if s="`system-control find dnscache`"
+then
+	set_if_unset dnscache IPSEND 0.0.0.0
+	set_if_unset dnscache IP 127.0.0.1
+#	set_if_unset dnscache DATALIMIT 3000000
+	set_if_unset dnscache CACHESIZE 1000000
+	set_if_unset dnscache ROOT "root"
 
-test -r "$s/service/seed" || dd if=/dev/urandom of="$s/service/seed" bs=128 count=1
-test -r "$s/service/root/servers/@" || echo '127.53.0.1' > "$s/service/root/servers/@"
-dir_not_empty "$s/service/root/ip" || touch "$s/service/root/ip/127.0.0.1"
+	system-control print-service-env dnscache >> "$3"
+
+	test -r "$s/service/seed" || dd if=/dev/urandom of="$s/service/seed" bs=128 count=1
+	test -r "$s/service/root/servers/@" || echo '127.53.0.1' > "$s/service/root/servers/@"
+	dir_not_empty "$s/service/root/ip" || touch "$s/service/root/ip/127.0.0.1"
+fi
+
+lr="/var/local/sv/"
+e="--no-systemd-quirks --escape-instance --bundle-root"
+
+list_network_addresses |
+while read -r i
+do
+	test -z "$i" && continue
+	service="dnscache@$i"
+	s="$lr/${service}"
+
+	system-control convert-systemd-units $e "$lr/" "./${service}.socket"
+	system-control preset "${service}"
+	rm -f -- "${s}/log"
+	ln -s -- "../../../sv/cyclog@dnscache" "${s}/log"
+
+	install -d -m 0755 "${s}/service/env"
+	install -d -m 0755 "${s}/service/root"
+	install -d -m 0755 "${s}/service/root/ip"
+	install -d -m 0755 "${s}/service/root/servers"
+	test -r "${s}/service/seed" || dd if=/dev/urandom of="${s}/service/seed" bs=128 count=1
+	test -r "${s}/service/root/servers/@" || echo '127.53.0.1' > "${s}/service/root/servers/@"
+	dir_not_empty "${s}/service/root/ip" || touch "${s}/service/root/ip/127.0.0.1"
+	set_if_unset "${s}/" IPSEND 0.0.0.0
+	set_if_unset "${s}/" CACHESIZE 1000000
+	set_if_unset "${s}/" ROOT "root"
+
+	if system-control is-enabled "${s}"
+	then
+		echo >> "$3" on "${service}"
+	else
+		echo >> "$3" off "${service}"
+	fi
+done
